Fix memory leak in disassembler
Reproduce by simply running the default flash Q program already
in memory and turning up the speed to the max.

The process memory size grows by about 2MB per second.

Not completely sure that this is the correct fix, but it seems to
work.

See:
https://doc.qt.io/qt-5/qplaintextedit.html#using-qplaintextedit-as-a-display-widget

for a discussion that suggests this fix is the right approach for using
this widget as a log viewer.  I didn't see any way to delete blocks as
lines/paragraphs are appended.